requirements.txt
flask
langchain
langchain-openai
python-dotenv
duckduckgo-search>=6.0
langchain-community
pytest
flake8
langgraph
langgraph-checkpoint
ddgs
mypy
fastapi
pydantic
uvicorn
httpx
jinja2
